smallbasket fn_801816F8 99.19->99.31: drop spurious (s16) switch cast (mode is int from u8 data[0x1e]/literals; extsh elided, sth-truncate path matches target)

+++ src/main/dll/dll_0104_smallbasket.c
@@ -211,7 +211,7 @@ int fn_801816F8(u8* obj, u8* player, u8* dataIn)
     }
 
     vel = gSmallBasketHitVelocity;
-    switch ((s16)mode)
+    switch (mode)
     {
     case 1:
         setup = Obj_AllocObjectSetup(0x24, 0x3d3);
